refactor: code conventions II
