+++ ee/packages/amplication-git-pull-service/src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { GitPullEventModule } from "./git-pull-event/git-pull-event.module";
 import { HealthModule } from "./health/health.module";
-import { SecretsManagerModule } from "./providers/secrets/secretsManager.module";
+import { SecretsManagerModule } from "./secrets/secretsManager.module";
 import { MorganModule } from "nest-morgan";
 import { ConfigModule } from "@nestjs/config";
 import { RootWinstonModule } from "./winston/root-winston.module";
