--- ee/packages/amplication-git-pull-service/src/constants/logger-messages.ts
@@ -1,18 +1,0 @@
-// TODO: should be an enum
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export const LoggerMessages = {
-  log: {
-    GENERATE_OCTOKIT_ACCESS_TOKEN: "generate access token with octokit",
-    NEW_GIT_PULL_EVENT_RECORD_CREATED: "create new git pull event record on DB",
-    FOUND_PREVIOUS_READY_COMMIT: "found previous ready commit",
-    CLONE_SUCCESS: "clone successfully",
-    PULL_SUCCESS: "pull successfully",
-    COPY_SUCCESS: "copy successfully",
-    DELETE_SUCCESSFULLY:
-      "update record status to Deleted and delete from storage",
-  },
-  error: {
-    CATCH_ERROR_MESSAGE: "error from:",
-  },
-  debug: {},
-};
