--- ee/packages/amplication-git-pull-service/src/enums/git-fetch-type.enum.ts
@@ -1,4 +1,0 @@
-export enum GitFetchTypeEnum {
-  Clone = "clone",
-  Pull = "pull",
-}
