--- ee/packages/amplication-git-pull-service/src/enums/git-provider.enum.ts
@@ -1,3 +1,0 @@
-export enum GitProviderEnum {
-  Github = "Github",
-}
