--- ee/packages/amplication-git-pull-service/src/enums/git-pull-event-status.enum.ts
@@ -1,5 +1,0 @@
-export enum EnumGitPullEventStatus {
-  Created = "Created",
-  Ready = "Ready",
-  Deleted = "Deleted",
-}
