--- ee/packages/amplication-git-pull-service/src/enums/index.ts
@@ -1,3 +1,0 @@
-export { GitFetchTypeEnum } from "./git-fetch-type.enum";
-export { GitProviderEnum } from "./git-provider.enum";
-export { EnumGitPullEventStatus } from "./git-pull-event-status.enum";
