--- ee/packages/amplication-git-pull-service/src/errors/custom-error.ts
@@ -1,5 +1,0 @@
-export class CustomError extends Error {
-  constructor(message: string, cause?: any) {
-    super(`${message}: ${cause}`);
-  }
-}
