--- ee/packages/amplication-git-pull-service/src/errors/simple-custom-error.ts
@@ -1,1 +1,0 @@
-export class SimpleCustomError extends Error {}
