+++ ee/packages/amplication-git-pull-service/src/git-pull-event/git-host-provider.service.ts
@@ -1,10 +1,9 @@
 import { Injectable } from "@nestjs/common";
-import { GitProvider } from "../../interfaces";
+import { GitProvider } from "./types";
 import { App } from "octokit";
-import { convertToNumber } from "../../utils/convert-to-number";
+import { convertToNumber } from "../utils/convert-to-number";
 import { ConfigService } from "@nestjs/config";
-import { CustomError } from "../../errors/custom-error";
-import { ErrorMessages } from "../../constants/error-messages";
+import { ErrorMessages } from "./constants";
 
 const GITHUB_APP_APP_ID_VAR = "GITHUB_APP_APP_ID";
 const GITHUB_APP_PRIVATE_KEY_VAR = "GITHUB_APP_PRIVATE_KEY";
@@ -42,7 +41,7 @@ export class GitHostProviderService implements GitProvider {
       );
       return data.token;
     } catch (err) {
-      throw new CustomError(ErrorMessages.ACCESS_TOKEN_ERROR, err);
+      throw new Error(ErrorMessages.ACCESS_TOKEN_ERROR, { cause: err });
     }
   }
 }
