+++ ee/packages/amplication-git-pull-service/src/git-pull-event/git-pull-event.module.ts
@@ -1,17 +1,17 @@
 import { Module } from "@nestjs/common";
 import { GitPullEventService } from "./git-pull-event.service";
 import { GitPullEventController } from "./git-pull-event.controller";
-import { PrismaService } from "../prisma.service";
 import { GitPullEventRepository } from "./git-pull-event.repository";
-import { StorageService } from "../providers/storage/storage.service";
-import { GitHostProviderService } from "../providers/git/git-host-provider.service";
-import { GitClientService } from "../providers/git/git-client.service";
-import { GitHostProviderFactory } from "../providers/git/git-host-provider-factory";
+import { StorageService } from "./storage.service";
+import { GitHostProviderService } from "./git-host-provider.service";
+import { GitClientService } from "./git-client.service";
+import { GitHostProviderFactory } from "./git-host-provider-factory";
+import { PrismaModule } from "../prisma/prisma.module";
 
 @Module({
+  imports: [PrismaModule],
   controllers: [GitPullEventController],
   providers: [
-    PrismaService,
     GitPullEventService,
     GitHostProviderService,
     {
