+++ ee/packages/amplication-git-pull-service/src/health/base/health.service.base.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { PrismaService } from "../../prisma.service";
+import { PrismaService } from "../../prisma/prisma.service";
 
 @Injectable()
 export class HealthServiceBase {
