+++ ee/packages/amplication-git-pull-service/src/health/health.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../prisma/prisma.service";
 import { HealthController } from "./health.controller";
 import { HealthService } from "./health.service";
