+++ ee/packages/amplication-git-pull-service/src/health/health.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@nestjs/common";
-import { PrismaService } from "../prisma.service";
+import { PrismaService } from "../prisma/prisma.service";
 import { HealthServiceBase } from "./base/health.service.base";
 
 @Injectable()
