--- ee/packages/amplication-git-pull-service/src/interfaces/git-client.interface.ts
@@ -1,14 +1,0 @@
-import { PushEventMessage } from "./push-event-message";
-
-export interface GitClient {
-  clone: (
-    pushEventMessage: PushEventMessage,
-    baseDir: string,
-    accessToken: string
-  ) => Promise<void>;
-  pull: (
-    pushEventMessage: PushEventMessage,
-    baseDir: string,
-    accessToken: string
-  ) => Promise<void>;
-}
