--- ee/packages/amplication-git-pull-service/src/interfaces/git-host-provider-factory.interface.ts
@@ -1,6 +1,0 @@
-import { GitProvider } from "./git-provider.interface";
-import { GitProviderEnum } from "../enums";
-
-export interface GitHostProviderFactory {
-  getHostProvider: (provider: GitProviderEnum) => GitProvider;
-}
