--- ee/packages/amplication-git-pull-service/src/interfaces/git-provider.interface.ts
@@ -1,3 +1,0 @@
-export interface GitProvider {
-  createInstallationAccessToken: (installationId: string) => Promise<string>;
-}
