--- ee/packages/amplication-git-pull-service/src/interfaces/git-pull-event-repository.interface.ts
@@ -1,15 +1,0 @@
-import { GitProviderEnum, EnumGitPullEventStatus } from "../enums";
-import { EventData } from "./event-data";
-
-export interface GitPullEventRepository {
-  create: (eventData: EventData) => Promise<{ id: bigint }>;
-  update: (id: bigint, status: EnumGitPullEventStatus) => Promise<boolean>;
-  findByPreviousReadyCommit: (
-    provider: keyof typeof GitProviderEnum,
-    repositoryOwner: string,
-    repositoryName: string,
-    branch: string,
-    pushedAt: Date,
-    skip: number
-  ) => Promise<EventData | undefined>;
-}
