--- ee/packages/amplication-git-pull-service/src/interfaces/git-pull-event.interface.ts
@@ -1,5 +1,0 @@
-import { PushEventMessage } from "./push-event-message";
-
-export interface GitPullEvent {
-  handlePushEvent: (pushEventMessage: PushEventMessage) => Promise<void>;
-}
