--- ee/packages/amplication-git-pull-service/src/interfaces/index.ts
@@ -1,8 +1,0 @@
-export { EventData } from "./event-data";
-export { GitClient } from "./git-client.interface";
-export { GitHostProviderFactory } from "./git-host-provider-factory.interface";
-export { GitProvider } from "./git-provider.interface";
-export { GitPullEventRepository } from "./git-pull-event-repository.interface";
-export { GitPullEvent } from "./git-pull-event.interface";
-export { PushEventMessage } from "./push-event-message";
-export { Storage } from "./storage.interface";
