--- ee/packages/amplication-git-pull-service/src/interfaces/storage.interface.ts
@@ -1,4 +1,0 @@
-export interface Storage {
-  copyDir: (srcDir: string, destDir: string) => Promise<void>;
-  deleteDir: (dir: string) => void;
-}
