+++ ee/packages/amplication-git-pull-service/src/prisma/prisma.module.ts
@@ -1,0 +1,7 @@
+import { Module } from "@nestjs/common";
+import { PrismaService } from "./prisma.service";
+
+@Module({
+  providers: [PrismaService],
+})
+export class PrismaModule {}
