--- ee/packages/amplication-git-pull-service/src/providers/secrets/secretsManager.module.ts
@@ -1,8 +1,0 @@
-import { Module } from "@nestjs/common";
-import { SecretsManagerService } from "./secretsManager.service";
-
-@Module({
-  providers: [SecretsManagerService],
-  exports: [SecretsManagerService],
-})
-export class SecretsManagerModule {}
