--- ee/packages/amplication-git-pull-service/src/providers/secrets/secretsManager.service.ts
@@ -1,10 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { ConfigService } from "@nestjs/config";
-import { SecretsManagerServiceBase } from "./base/secretsManager.service.base";
-
-@Injectable()
-export class SecretsManagerService extends SecretsManagerServiceBase {
-  constructor(protected readonly configService: ConfigService) {
-    super(configService);
-  }
-}
