+++ ee/packages/amplication-git-pull-service/src/utils/convert-to-number.ts
@@ -1,9 +1,7 @@
-import { SimpleCustomError } from "../errors/simple-custom-error";
-
 export const convertToNumber = (value: string): number => {
   const result = parseInt(value);
   if (isNaN(result)) {
-    throw new SimpleCustomError(
+    throw new Error(
       "error from convertToNumber() => GitHub App installation identifier is invalid"
     );
   }
