+++ ee/packages/amplication-git-pull-service/src/variables.ts
